requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
requests
pydantic
scipy
soundfile
torch
transformers
python-dotenv